+++ NEWS.md
@@ -1,4 +1,19 @@
 
+# hclust1d 0.1.0
+
+- Added a vignette on replacing `stats::hclust` with `hclust1d`
+- Added a getting started vignette
+- Added citation info and added `.onAttach` message with this citation info
+- Added a listing of all currently supported distance methods in `supported_methods`
+- Added a listing of all currently supported distance methods in `supported_dist.methods`
+- Added `complete` linkage support
+- Added `average` linkage support
+- Added `centroid` (*UPGMC*) and `median` (*weighted centroid, WPGMC*) linkage support
+- Added `true_median` linkage support
+- Added `ward.D` and `ward.D2` linkage support
+- Added `mcquitty` (*WPGMC*) linkage support
+- Added a heap implementation of priority queue with key update by id
+
 # hclust1d 0.0.1
 
-- The initial version with single linkage, distance support and unit tests
+- The initial version with `single` linkage support, distance support and unit tests